Bug fixed in validateInput and check functions.
A bug resulted in a failed loop when a misspelling entered the check
more than once. This was fixed by changing the check to look for a false
return from the validateInput function. Another bug came up when the
loop then failed to close after one nonconformity. The addtion of
resetting the input back to true at line 39 by resetting input to
lowercase, and running the check again.